Add timeout for reading the range from the sensor
It happens that the reading blocks indefinitely in one of the `while`
loops that wait for a certain bit to be set. It seems there is no
problem with just continuing the read procedure if the bits are not what
they're supposed to be. The actual range value returned is probably
outdated which would be the only caveat. This blocking only occurs once
(at least for me), i.e. having the timeout when the blocking occurs
results in at most one outdated value. All subsequent reads can be
performed no problem.
THe default behavior is that there is no timeout to be backwards
compatible.